fix(camera): address PR #253 review feedback
Schema/tooling:
- De-version the SBMD schema surface: rename to sbmd-spec-schema.json and
  validate_sbmd_specs.py, drop the schema version from filenames and docs, and
  track versions in schema/CHANGELOG.md. schemaVersion is still validated
  against the schema's expected version.

Matter data model:
- Fix the WebRTCTransportRequestor ClusterRevision attribute default (1 -> 2) in
  the ZAP so the reported revision matches the cluster definition, and
  regenerate barton-library.matter from the ZAP via the SDK generator.
- Matter::HasCaseOperateAclEntryForCluster: treat ACL entries with no targets as
  granting all clusters (CHIP ACL semantics), avoiding redundant ACL entries.

camera.sbmd.js:
- Guard parseSessions() null in the incoming Offer/End handlers, reject empty
  SDP in sendProvideAnswer, document stream() returns a JSON string.
- takePicture takes the sessions transient-data supplement; deviceClassVersion
  2; named constants for stream-usage and end-reason; findStreamingSessionId
  helper; incomingEnd -> incomingEndSession; drop leftover DIAG-STEP debug;
  single-return pickNegotiationRole; fix contradictory entryPoint comment.

Tests:
- Extract the shared SBMD test harness (C stubs, runtime setup, JS/TLV helpers,
  supplement plumbing, ParsedResult assertions) into SbmdDriverTestBase and
  SbmdDriverTestSupport.cpp; SbmdCameraWebrtcTest and SbmdHandlerInvokerTest now
  derive from it. Add SessionsJson helper and split the ICE-candidates literal.
