+++ core/deviceDrivers/matter/sbmd/schema/CHANGELOG.md
@@ -1,7 +1,16 @@
 # SBMD Schema Changelog
 
+The schema lives in a single file, `sbmd-spec-schema.json`, whose version is not
+encoded in the filename. This changelog records the schema version history, and
+each spec's `schemaVersion` field is validated against the schema's expected
+version. Each versioned heading below describes the changes introduced by that
+version.
+
 ## v5.0
 
+- The schema version is no longer encoded in the schema filename
+  (`sbmd-spec-schema.json`) or in docs; `schemaVersion` is still validated
+  against the schema's expected version
 - Breaking: attribute/event/command handlers must bind via `aliases`. The inline
   `clusterId` + `attributeId`/`attributeIds` (and `eventId`/`eventIds`,
   `commandId`/`commandIds`) binding form has been removed; declare an alias in the
